ocr_api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart
pillow>=10.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
numpy>=1.24.0